Prevent double layering of generated module name
In the normal Rust module system, a module is represented in the
filesystem by a file of the same name. With the previous setup,
the generated code was saved in `<target_mod>.rs`, with that
file containing another `pub mod <target_mod> { .. }` inside.
Here we remove the outer containing module from the generated code
and instead add it in the cdp crate root, where the generated code
is included, to keep the same paths.