fix: Fix Clippy issues
This commit fixes Clippy errors. The main issue was the new rule to use identifiers in format strings, fixed with `clippy --fix`. It also fixes an example to use `while let Some(_)` instead of a `loop { match {...}}`.

Finally, it also fixes the error message the recently added `CdpError::LaunchIo` variant.