feat: Expose async `kill` and `wait`
This commit expands on the changes from mattsse/chromiumoxide#128 (async child process) to provide async support for `kill` and `wait` calls, used to manage the spawned chromium instance.

The existing `kill` and `wait` methods were renamed to `kill_sync` and `wait_sync`. New asynchronous implementations of `kill` and `wait` are provided. The methods from `async_process::Child` were renamed for consistency (async by default, `_sync` suffix for blocking variants).
The changes above are breaking, but they take advantage of the fact that switching to async processes was already a breaking change and that they were not published yet (so it's still a single breaking change).

In addition to these changes, the documentation was updated to provide more information and cross-link related methods. A couple helper methods were added to retrieve the concret child process implementation.