fix: Wait on child process if `launch` fails
If an error occurs in `Browser::launch` after the browser process is spawned but before the `Browser` is fully constructed, the child process is not `wait`-ed on. This can cause a chain reaction were a timeout causes a zombie process consuming resources, and slowing the system down; causing even more errors.

After `Browser` is created, the `Drop` implementation takes care of waiting on the child process.

This commit fixes this error by wrapping all faillible steps of `Browser::launch` in an inner function, and then properly cleaning-up the child process if any error occurs.